requirements.txt
annotated-types==0.7.0
anyio==4.7.0
click==8.1.7
cffi==1.17.1
cryptography==43.0.3
pycparser==2.22
PyJWT==2.6.0
Deprecated==1.2.15
fastapi==0.115.6
h11==0.14.0
idna==3.10
limits==3.14.1
packaging==24.2
passlib==1.7.4
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
python-dotenv==1.0.1
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==1.4.45
typing_extensions==4.12.2
uvicorn==0.32.1
wrapt==1.17.0
bcrypt==4.2.1
loguru==0.7.3
pymysql==1.1.1
